+++ src/components/ExportBytesDialog.vue
@@ -1,5 +1,5 @@
 /** * VULNEX -Bytes Revealer- * * File: ExportBytesDialog.vue * Author: Simon Roses Femerling *
-Created: 2025-09-24 * Last Modified: 2025-09-27 * Version: 0.3 * License: Apache-2.0 * Copyright (c)
+Created: 2025-09-24 * Last Modified: 2025-09-27 * Version: 0.4 * License: Apache-2.0 * Copyright (c)
 2025 VULNEX. All rights reserved. * https://www.vulnex.com */
 
 <template>